Move IO functions to new file
