+++ source/utils/RandomUtils.h
@@ -1,7 +1,7 @@
 // ----------------------------------------------------------------------------
 // nexus | RandomUtils.h
 //
-// Commonly used generic functions.
+// Commonly used functions to generate random quantities.
 //
 // The NEXT Collaboration
 // ----------------------------------------------------------------------------
@@ -35,27 +35,6 @@ namespace nexus {
     
     /// Check if the sampled value is out of bounds, max check only
     G4bool CheckOutOfBoundMax(G4double max, G4double val);
-
-    /// Read in the 1d histogram stored in a csv file
-    void LoadHistData1D(std::string filename, std::vector<G4double> &value,
-                        std::vector<G4double> &x,
-                        std::vector<G4double> &x_smear);
-    
-    /// Read in the 2d histogram stored in a csv file
-    void LoadHistData2D(std::string filename, std::vector<G4double> &value,
-                        std::vector<G4double> &x, std::vector<G4double> &y,
-                        std::vector<G4double> &x_smear,
-                        std::vector<G4double> &y_smear);
-    
-    /// Read in the 3d histogram stored in a csv file
-    void LoadHistData3D(std::string filename, std::vector<G4double> &value,
-                        std::vector<G4double> &x, std::vector<G4double> &y, std::vector<G4double> &z,
-                        std::vector<G4double> &x_smear, std::vector<G4double> &y_smear, std::vector<G4double> &z_smear);
-
-
-    // Check if a loaded variable from a csv is in the desired range
-    // Header name is the string in the file that identifies the bins you want to check. e.g. energy, azimuth, zenith
-    void CheckVarBounds(std::string filename, G4double var_min, G4double var_max, std::string HeaderName);
 
 
 }
